Feat: isFetcing true 일 시, Loading 아이콘 적용 #8
